Install matter-idl-lint as a part of matter-idl package
