+++ scripts/py_matter_idl/setup.cfg
@@ -33,7 +33,8 @@ install_requires =
 [options.entry_points]
 console_scripts =
     matter-data-model-xml-parser = matter.idl.data_model_xml:main
-    matter-idl-lint-parser = matter.idl.lint:main
+    matter-idl-lint = matter.idl.lint:main
+    matter-idl-lint-parser = matter.idl.lint:parser
     matter-idl-parser = matter.idl.matter_idl_parser:main
     matter-idl-check-backward-compatibility = matter.idl.backwards_compatibility:main
     matter-zapxml-parser = matter.idl.zapxml:main
